Added missing class members
Still not sure if certain members should be initialized outside of the
class and passed in through a setter or in the constructor or intialize
them in the constructor. Both have their pros and cons. Any opinion. I'm
going to go with the first option for now. However, I need to identify
which one's will require that.